requirements.txt
torch
torchvision
tensorboardX
tensorflow
pyyaml
torchscan